+++ Mantids30/Net_Sockets/socket.h
@@ -230,7 +230,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual int partialRead(void * data, const uint32_t & datalen);
+    virtual ssize_t partialRead(void * data, const uint32_t & datalen);
     /**
      * Write a data block to the socket
      * note that this haves some limitations. some systems can only send 4k at time.
@@ -239,7 +239,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual int partialWrite(const void *data, const uint32_t & datalen);
+    virtual ssize_t partialWrite(const void *data, const uint32_t & datalen);
 
     /**
      * @brief iShutdown Internal protocol Shutdown
